+++ packages/dsw-tdk/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [3.20.1]
+
+### Fixed
+
+- Fix reported version using `--version`
+- Fix creation of document template draft if DSW instance uses different organization ID
+
 ## [3.20.0]
 
 ### Added
@@ -274,3 +281,4 @@ Initial DSW Template Development Kit (versioned as part of the [DSW platform](ht
 [3.19.1]: /../../tree/v3.19.1
 [3.19.2]: /../../tree/v3.19.2
 [3.20.0]: /../../tree/v3.20.0
+[3.20.1]: /../../tree/v3.20.1